@@ -3,10 +3,10 @@ version '1.0-SNAPSHOT'
 
 apply plugin: 'java'
 
-compileJava {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
-}
+//compileJava {
+//    sourceCompatibility = JavaVersion.VERSION_11
+//    targetCompatibility = JavaVersion.VERSION_11
+//}
 
 repositories {
     mavenCentral()
