@@ -1,6 +1,5 @@
 package com.vincentks.vic.game;
 
-public interface Event
-{
-  Item execute();
+public interface Event {
+	Item execute();
 }
