@@ -1,10 +1,9 @@
 package com.vincentks.vic.game;
 
-public interface Game
-{
-  CycleSummary cycle();
+public interface Game {
+	CycleSummary cycle();
 
-  void add(Actor actor, Item item);
+	void add(Actor actor, Item item);
 
-  void remove(Actor actor, Item item);
+	void remove(Actor actor, Item item);
 }
