@@ -3,9 +3,14 @@ package com.vincentks.vic.game;
 import java.util.UUID;
 
 // TODO add definition of each of the interfaces in this hierarchy
-public interface Item
-{
-  Effort buildEffort();
+public interface Item {
+	Effort buildEffort();
 
-  UUID getId();
+	UUID getId();
+
+	default boolean isCyclable() {
+		return false;
+	}
+
+	default boolean isDiffAware() { return false; }
 }
