@@ -2,9 +2,10 @@ package com.vincentks.vic.game;
 
 import java.util.Optional;
 
-public interface ItemProducer extends DiffAware
-{
-  Optional<Item> currentActivity();
-  void build(Item item);
-  Optional<Item> currentlyBuilding();
+public interface ItemProducer extends DiffAware {
+	Optional<Item> currentActivity();
+
+	void build(Item item);
+
+	Optional<Item> currentlyBuilding();
 }
