@@ -1,8 +1,7 @@
 package com.vincentks.vic.game;
 
-public interface LocationManager
-{
-  Location getLocation(MobileItem item);
+public interface LocationManager {
+	Location getLocation(MobileItem item);
 
-  void move(MobileItem item, Location location);
+	void move(MobileItem item, Location location);
 }
