@@ -14,64 +14,60 @@ import org.junit.Test;
 
 import com.vincentks.vic.game.util.MobileItemWithEffort;
 
-public class GameDiffWrapperTest
-{
-  @Test
-  public void testCycle_EmptyDiff() throws Exception
-  {
-    final CycleDiffCollector cycleDiffCollector = new CycleDiffCollector();
-    new GameDiffWrapper(
-        new GameImpl(),
-        NULL_ACTOR,
-        cycleDiffCollector
-    ).cycle();
+public class GameDiffWrapperTest {
+	@Test
+	public void testCycle_EmptyDiff() throws Exception {
+		final CycleDiffCollector cycleDiffCollector = new CycleDiffCollector();
+		new GameDiffWrapper(
+			new GameImpl(),
+			NULL_ACTOR,
+			cycleDiffCollector
+		).cycle();
 
-    assertThat(cycleDiffCollector.getDiffs(), emptyIterable());
-  }
+		assertThat(cycleDiffCollector.getDiffs(), emptyIterable());
+	}
 
-  @Test
-  public void testCycle_CityBuildingUnit() throws Exception
-  {
-    final CycleDiffCollector cycleDiffCollector = new CycleDiffCollector();
-    final Game game = new GameImpl();
-    final City city = EMPTY_CITY;
-    game.add(NULL_ACTOR, city);
-    final GameDiffWrapper wrapper = new GameDiffWrapper(
-        game,
-        NULL_ACTOR,
-        cycleDiffCollector
-    );
-    wrapper.cycle();
-    city.build(new MobileItemWithEffort(1));
-    wrapper.cycle();
+	@Test
+	public void testCycle_CityBuildingUnit() throws Exception {
+		final CycleDiffCollector cycleDiffCollector = new CycleDiffCollector();
+		final Game game = new GameImpl();
+		final City city = EMPTY_CITY;
+		game.add(NULL_ACTOR, city);
+		final GameDiffWrapper wrapper = new GameDiffWrapper(
+			game,
+			NULL_ACTOR,
+			cycleDiffCollector
+		);
+		wrapper.cycle();
+		city.build(new MobileItemWithEffort(1));
+		wrapper.cycle();
 
-    final CycleDiff cycleDiff = cycleDiffCollector.getDiffs().iterator().next();
-    assertThat(cycleDiff.getLevel(), is(NORMAL));
-    assertThat(cycleDiff.getMessage(), containsString("Created 1 new unit(s)"));
-  }
+		final CycleDiff cycleDiff = cycleDiffCollector.getDiffs().iterator().next();
+		assertThat(cycleDiff.getLevel(), is(NORMAL));
+		assertThat(cycleDiff.getMessage(), containsString("Created 1 new unit(s)"));
+	}
 
-  @Test
-  public void testCycle_CityPopulationEvolution() throws Exception
-  {
-    final CycleDiffCollector cycleDiffCollector = new CycleDiffCollector();
-    // TODO create a game builder?
-    final Game game = new GameImpl();
-    final City city = new CityBuilder()
-        .setName("Amsterdam")
-        .setLocation(new Location(DESERT))
-        .setPopulation(INITIAL_POPULATION)
-        .build();
-    game.add(NULL_ACTOR, city);
-    final GameDiffWrapper wrapper = new GameDiffWrapper(
-        game,
-        NULL_ACTOR,
-        cycleDiffCollector
-    );
-    wrapper.cycle();
-    wrapper.cycle();
+	@Test
+	public void testCycle_CityPopulationEvolution() throws Exception {
+		final CycleDiffCollector cycleDiffCollector = new CycleDiffCollector();
+		// TODO create a game builder?
+		final Game game = new GameImpl();
+		final City city = new CityBuilder()
+			.setName("Amsterdam")
+			.setLocation(new Location(DESERT))
+			.setPopulation(INITIAL_POPULATION)
+			.build();
+		game.add(NULL_ACTOR, city);
+		final GameDiffWrapper wrapper = new GameDiffWrapper(
+			game,
+			NULL_ACTOR,
+			cycleDiffCollector
+		);
+		wrapper.cycle();
+		wrapper.cycle();
 
-    final CycleDiff cycleDiff = cycleDiffCollector.getDiffs().iterator().next();
-    assertThat(cycleDiff.getLevel(), is(NORMAL));
-    assertThat(cycleDiff.getMessage(), containsString("Population changed to 1210 (+110)"));
-  }
+		final CycleDiff cycleDiff = cycleDiffCollector.getDiffs().iterator().next();
+		assertThat(cycleDiff.getLevel(), is(NORMAL));
+		assertThat(cycleDiff.getMessage(), containsString("Population changed to 1210 (+110)"));
+	}
 }
