@@ -11,61 +11,55 @@ import org.junit.Test;
 
 import com.vincentks.vic.game.util.NullActor;
 
-public class GameImplTest
-{
+public class GameImplTest {
 
-  @Test
-  public void testTurn_EmptyGame() throws Exception
-  {
-    CycleSummary cycleSummary = new GameImpl().cycle();
+	@Test
+	public void testTurn_EmptyGame() throws Exception {
+		CycleSummary cycleSummary = new GameImpl().cycle();
 
-    assertEquals(0, cycleSummary.numberOfItems());
-  }
+		assertEquals(0, cycleSummary.numberOfItems());
+	}
 
-  @Test
-  public void testCycle_TrackNumberOfTurns() throws Exception
-  {
-    final CycleSummary cycleSummary = new GameImpl().cycle();
+	@Test
+	public void testCycle_TrackNumberOfTurns() throws Exception {
+		final CycleSummary cycleSummary = new GameImpl().cycle();
 
-    assertThat(cycleSummary.getId(), is(1));
-  }
+		assertThat(cycleSummary.getId(), is(1));
+	}
 
-  @Test
-  public void testCycle_ItemGeneratesOneActionUponTurn() throws Exception
-  {
-    final Game game = new GameImpl();
-    game.add(NULL_ACTOR, NULL_DIFF_AWARE_ITEM);
+	@Test
+	public void testCycle_ItemGeneratesOneActionUponTurn() throws Exception {
+		final Game game = new GameImpl();
+		game.add(NULL_ACTOR, NULL_DIFF_AWARE_ITEM);
 
-    final CycleSummary cycleSummary = game.cycle();
+		final CycleSummary cycleSummary = game.cycle();
 
-    assertEquals(1, cycleSummary.numberOfItems());
-  }
+		assertEquals(1, cycleSummary.numberOfItems());
+	}
 
-  @Test
-  public void testCycle_SegregateSummaryByOwner() throws Exception
-  {
-    final Game game = new GameImpl();
-    final Actor actor = new NullActor();
-    game.add(actor, NULL_DIFF_AWARE_ITEM);
+	@Test
+	public void testCycle_SegregateSummaryByOwner() throws Exception {
+		final Game game = new GameImpl();
+		final Actor actor = new NullActor();
+		game.add(actor, NULL_DIFF_AWARE_ITEM);
 
-    game.add(new NullActor(), NULL_DIFF_AWARE_ITEM);
-    game.add(new NullActor(), NULL_DIFF_AWARE_ITEM);
+		game.add(new NullActor(), NULL_DIFF_AWARE_ITEM);
+		game.add(new NullActor(), NULL_DIFF_AWARE_ITEM);
 
-    final CycleSummary cycleSummary = game.cycle();
+		final CycleSummary cycleSummary = game.cycle();
 
-    assertEquals(3, cycleSummary.numberOfItems());
-    assertEquals(1, cycleSummary.itemsFor(actor).numberOfItems());
-  }
+		assertEquals(3, cycleSummary.numberOfItems());
+		assertEquals(1, cycleSummary.itemsFor(actor).numberOfItems());
+	}
 
-  @Test
-  public void testCycle_ShouldYieldNewItems() throws Exception
-  {
-    Item preTurnItem = EMPTY_CITY;
-    final Game game = new GameImpl();
-    game.add(NULL_ACTOR, preTurnItem);
+	@Test
+	public void testCycle_ShouldYieldNewItems() throws Exception {
+		Item preTurnItem = EMPTY_CITY;
+		final Game game = new GameImpl();
+		game.add(NULL_ACTOR, preTurnItem);
 
-    final CycleSummary cycleSummary = game.cycle();
+		final CycleSummary cycleSummary = game.cycle();
 
-    assertThat(cycleSummary.contains(preTurnItem), is(false));
-  }
+		assertThat(cycleSummary.contains(preTurnItem), is(false));
+	}
 }
