@@ -2,6 +2,5 @@ package com.vincentks.vic.game.util;
 
 import com.vincentks.vic.game.Actor;
 
-public class NullActor implements Actor
-{
+public class NullActor implements Actor {
 }
