@@ -10,29 +10,24 @@ import com.vincentks.vic.game.DiffAware;
 import com.vincentks.vic.game.Effort;
 import com.vincentks.vic.game.Item;
 
-public class NullDiffAwareItem implements DiffAware
-{
-  @Override
-  public Collection<CycleDiff> diff(Item itemInPreviousCycle)
-  {
-    return Collections.emptyList();
-  }
+public class NullDiffAwareItem implements DiffAware {
+	@Override
+	public Collection<CycleDiff> diff(Item itemInPreviousCycle) {
+		return Collections.emptyList();
+	}
 
-  @Override
-  public Cyclable cycle()
-  {
-    return this;
-  }
+	@Override
+	public Cyclable cycle() {
+		return this;
+	}
 
-  @Override
-  public Effort buildEffort()
-  {
-    return Effort.noEffort;
-  }
+	@Override
+	public Effort buildEffort() {
+		return Effort.noEffort;
+	}
 
-  @Override
-  public UUID getId()
-  {
-    return UUID.randomUUID();
-  }
+	@Override
+	public UUID getId() {
+		return UUID.randomUUID();
+	}
 }
